+++ tests/test_stac_api_workspace.py
@@ -7,6 +7,7 @@ import pytest
 import responses
 from mock import ANY, MagicMock
 from openeo_driver.testing import DictSubSet
+from openeo_driver.constants import ITEM_LINK_PROPERTY
 from pystac import Collection, Extent, SpatialExtent, TemporalExtent, Item, Asset, Link, RelType
 
 from openeogeotrellis.workspace import StacApiWorkspace
@@ -377,6 +378,46 @@ def test_vito_stac_api_workspace_helper(tmp_path, requests_mock, mock_s3_bucket)
     assert object_asset_object_metadata["mtime"] == "1756477082123456789"
 
 
+def test_merge_with_auxiliary_file(requests_mock, tmp_path):
+    asset_path = Path("/path") / "to" / "asset.tif"
+    collection = _collection(
+        root_path=tmp_path / "collection",
+        collection_id="collection",
+        asset_hrefs=[str(asset_path)],
+        auxiliary_href="/path/to/auxiliary_file",
+    )
+
+    export_link_mock = MagicMock(wraps=_export_link)
+
+    stac_api_workspace = StacApiWorkspace(
+        root_url="https://stacapi.test",
+        export_asset=_export_asset,
+        asset_alternate_id="file",
+        export_link=export_link_mock,
+    )
+
+    target = PurePath("new_collection")
+
+    _mock_stac_api_root_catalog(requests_mock, stac_api_workspace.root_url)
+    requests_mock.get(f"{stac_api_workspace.root_url}/collections/{target.name}", status_code=404)
+    create_collection_mock = requests_mock.post(f"{stac_api_workspace.root_url}/collections")
+    create_item_mock = requests_mock.post(f"{stac_api_workspace.root_url}/collections/{target.name}/items")
+
+    stac_api_workspace.merge(collection, target)
+
+    assert create_collection_mock.called_once
+
+    assert create_item_mock.called_once
+    assert create_item_mock.last_request.json()["links"] == [
+        {
+            "rel": "aux",
+            "href": "/path/to/auxiliary_file",
+        }
+    ]
+
+    export_link_mock.assert_called_once_with(ANY, target, False)
+
+
 def _mock_stac_api_root_catalog(requests_mock, root_url: str):
     # STAC API root catalog with "conformsTo" for pystac_client
     requests_mock.get(
@@ -396,11 +437,20 @@ def _mock_stac_api_root_catalog(requests_mock, root_url: str):
 
 
 def _export_asset(asset: Asset, merge: PurePath, relative_asset_path: PurePath, remove_original: bool) -> str:
+    assert isinstance(asset, Asset)
     assert isinstance(merge, PurePath)
     assert isinstance(relative_asset_path, PurePath)
     assert isinstance(remove_original, bool)
     # actual copying behaviour is the responsibility of the workspace creator
     return asset.get_absolute_href()
+
+
+def _export_link(link: Link, merge: PurePath, remove_original: bool) -> str:
+    assert isinstance(link, Link)
+    assert isinstance(merge, PurePath)
+    assert isinstance(remove_original, bool)
+    # actual copying behaviour is the responsibility of the workspace creator
+    return link.get_absolute_href()
 
 
 def _asset_workspace_uris(collection: Collection, alternate_key: str) -> Dict[str, str]:
@@ -417,6 +467,7 @@ def _collection(
     asset_hrefs: List[str] = None,
     spatial_extent: SpatialExtent = SpatialExtent([[-180, -90, 180, 90]]),
     temporal_extent: TemporalExtent = TemporalExtent([[None, None]]),
+    auxiliary_href: str = None,
 ) -> Collection:
     if asset_hrefs is None:
         asset_hrefs = []
@@ -434,9 +485,16 @@ def _collection(
         item_id = asset_key = asset_filename
 
         item = Item(id=item_id, geometry=None, bbox=None, datetime=dt.datetime.now(dt.timezone.utc), properties={})
-        asset = Asset(href=asset_href)
 
+        asset = Asset(href=asset_href)
         item.add_asset(asset_key, asset)
+
+        if auxiliary_href is not None:
+            auxiliary_link = Link(
+                rel="aux", target=auxiliary_href, extra_fields={ITEM_LINK_PROPERTY.EXPOSE_AUXILIARY: True}
+            )
+            item.add_link(auxiliary_link)
+
         collection.add_item(item)
 
         collection.add_link(Link(rel=RelType.DERIVED_FROM, target=f"https://src.test/{asset_filename}"))